requirements.txt
# Core LangChain and AI
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.70
langchain-ollama==0.3.5
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8

# Vector Database and Embeddings
chromadb==0.5.23
sentence-transformers==5.0.0
langchain-huggingface

# OpenAI and Azure
openai==1.97.0

# Voice and Audio
pyttsx3==2.99
SpeechRecognition==3.14.3
PyAudio==0.2.14

# API Requests
requests==2.32.4
httpx==0.28.1

# Environment and Configuration
python-dotenv==1.1.1
PyYAML==6.0.2

# Document Processing - simplified to avoid conflicts
# unstructured==0.16.8

# Data Processing - compatible version
numpy>=1.22.5,<2.0.0
pandas>=2.0.0

# Utility
click==8.1.8
typing_extensions>=4.0.0
pydantic>=2.0.0
tqdm>=4.60.0
pytz>=2020.1

# Ollama Client
ollama==0.5.1

# Additional dependencies
aiofiles==24.1.0
fastapi==0.116.1
uvicorn==0.35.0
